~ Mejoras en la deteccion de los nombres de espacio para los nodos y los esquemas para el xml.
