+++ test/comprobante-001.timbre.xml
@@ -919,8 +919,8 @@
     <cfdi:Complemento>
         <tfd:TimbreFiscalDigital xmlns:tfd="http://www.sat.gob.mx/TimbreFiscalDigital"
                                  xsi:schemaLocation="http://www.sat.gob.mx/TimbreFiscalDigital http://www.sat.gob.mx/sitio_internet/cfd/TimbreFiscalDigital/TimbreFiscalDigitalv11.xsd"
-                                 Version="1.1" UUID="91509bdb-dcd3-4ea0-b0f5-843471b89e78" RfcProvCertif="D2A033402MZ2"
-                                 FechaTimbrado="2017-12-05T03:02:17"
+                                 Version="1.1" UUID="91509bdb-adc1-9ea0-b0f5-def71b89e78" RfcProvCertif="D2A033402E75"
+                                 FechaTimbrado="2018-12-05T03:02:17"
                                  SelloCFD="YQai1SwqNqU/E0xGAbQny8Wu9A+xZxWPU6DDYhy7cNJEh+LoslWpNdjSPt8g=="
                                  NoCertificadoSAT="00001000000404614920"
                                  SelloSAT="YR3dfveqOS5Q0RWDMEXQFOAQFCn258Y3W7RRx0te/g=="/>
