--- test/xml-005-schema-innode.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<cfdi:Comprobante xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.sat.gob.mx/cfd/3 http://www.sat.gob.mx/sitio_internet/cfd/3/cfdv33.xsd" LugarExpedicion="08400" MetodoPago="PUE" TipoDeComprobante="I" Total="250.00" Moneda="MXN" Certificado="MIIGVzCCBD+gAwIBAgIUMDAwMDEwMDAwMDA0MDMyNTg3NDgwDQYJKoZIhvcNAQELBQAwggGyMTgwNgYDVQQDDC9BLkMuIGRlbCBTZXJ2aWNpbyBkZSBBZG1pbmlzdHJhY2nDs24gVHJpYnV0YXJpYTEvMC0GA1UECgwmU2VydmljaW8gZGUgQWRtaW5pc3RyYWNpw7NuIFRyaWJ1dGFyaWExODA2BgNVBAsML0FkbWluaXN0cmFjacOzbiBkZSBTZWd1cmlkYWQgZGUgbGEgSW5mb3JtYWNpw7NuMR8wHQYJKoZIhvcNAQkBFhBhY29kc0BzYXQuZ29iLm14MSYwJAYDVQQJDB1Bdi4gSGlkYWxnbyA3NywgQ29sLiBHdWVycmVybzEOMAwGA1UEEQwFMDYzMDAxCzAJBgNVBAYTAk1YMRkwFwYDVQQIDBBEaXN0cml0byBGZWRlcmFsMRQwEgYDVQQHDAtDdWF1aHTDqW1vYzEVMBMGA1UELRMMU0FUOTcwNzAxTk4zMV0wWwYJKoZIhvcNAQkCDE5SZXNwb25zYWJsZTogQWRtaW5pc3RyYWNpw7NuIENlbnRyYWwgZGUgU2VydmljaW9zIFRyaWJ1dGFyaW9zIGFsIENvbnRyaWJ1eWVudGUwHhcNMTYwODAxMTczMDQ5WhcNMjAwODAxMTczMDQ5WjCB9zEuMCwGA1UEAxMlU0VSVklDSU8gREUgQURNSU5JU1RSQUNJT04gVFJJQlVUQVJJQTEuMCwGA1UEKRMlU0VSVklDSU8gREUgQURNSU5JU1RSQUNJT04gVFJJQlVUQVJJQTEuMCwGA1UEChMlU0VSVklDSU8gREUgQURNSU5JU1RSQUNJT04gVFJJQlVUQVJJQTElMCMGA1UELRMcU0FUOTcwNzAxTk4zIC8gVUFMTDY3MDQwMUE5NTEeMBwGA1UEBRMVIC8gVUFMTDY3MDQwMUhERkdHUzA1MR4wHAYDVQQLExVNRUdBUEFDU0FUOTcwNzAxTk4zMDEwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCPO6e696P+vfbYmTUA/mEiA4/+csSSFOPxbbjW4+CihXeFldm+KuovQQwK13slgdlkmmTyFdCMD52tbytiTDXrRPATs4wPS5422IhntaJfZvBidVaBO19ky/60zkoJqo5/UrBR+ld3V6QpaP7XXUQuXp0sBRbPtn991cB/i5v316CNvnWViVTpFZrU7e10t4gAiAzf7xhcvMzB/+ZUA8S5XRqWff9PbHolsZ88G4b2/h1N/0AfQXfn9ieWLDpCh6uVj/YlhYhVlvSYdK9AOiGMycVn6tuf2LwyOR2yxLesnEoJAGe0Liwl75KXzwKiL7N8vPUsHJqcxpHJdgsab1mpAgMBAAGjHTAbMAwGA1UdEwEB/wQCMAAwCwYDVR0PBAQDAgbAMA0GCSqGSIb3DQEBCwUAA4ICAQBj8+nr+M1IvQExxuMxfHLBZDezlJOFT9aDLe9mSfoGbZg34JDJoOEDBTYwoQVvr91ncMqBfzQIo7wmF1mrockW1cHuva8ul5QmNHvwdGOre4FnZANmnsCIWVpkCaEq0LOS8E1qrp2MHZ7EThFElN0tPVArZ6ZW9Njj+e7lXfGignbayp0d1SEkuwhy0pH1TvsaTLsuskdbEQHAhMXTRzMUHgs7jeo80LtKNzkxocLpiSa6GnCeIKeyZDvzW7Vj3H6xbWg3gmBz/PUZNW+B/+Q0QNnq3veQ5QKJLz/Fnly695Nti7YZTOg+0xSdWEK8flUMoB0josF0OSlLtLvnZBp7NxbwlLWA0DBTe+s4OoUrm9Lw8HYJ/OJhizwM7W9ffWQLOa0yTQBbCrMljy7CaIeQyV2pn8DddO4RZX+FdA2jWpNecWPzGjx/xRIBuFPkzF6uSdrZY3qA0YcISgpXE7hH3Cf0KUIH68BzdoLcaMyBqLzmIfhnUnjsZLp+CoNnk9JGQCeDsHdnxOEYpnR8i9/rbEXMrGI4H+SbfwszHZKhc56ij76/8IorHdFKotvfnGu8MWhzz1cdtNr1SHlrLeqm8vqV9D4PjTeFpBSkQrPHDOrYsOXRvMS5QLDIYWUQYsAXm2LbLSgXN+upgZBbpuMTbxJ8C6f51ROpMxpFWgTvgQ==" SubTotal="215.52" NoCertificado="00001000000403258748" FormaPago="01" Sello="XpGru5ERb0yhJ8Fv3V64jwDAX8pJSGKEYTjCv6Y2j2jzNenmO2mntJ+93DE5dB+yTwhn5OHVZlWrEZV2OvZJ/MrP/IHdt5pZ8/ZTUs/Pm6K+BssJVUX3y20agzffLq4VzZq7zF1rxDjwN5lnsPMPlnqn0N4alC/KDSs1s1rV/GoZAh+pFtysR2a5AL17MfcjEGobu9U9MdEdHvmsc2W1ydbfnFBfVkyOqYRziHNu0m4929xxkZPrv3tse8zAZ6AxinH07YAMMQM5pv396RkqpFzsnp61M8K++wO/33XKxw1ybjRNJ4tRQ+QFL848x+8SyxTpF8vKxSNW8tg37f7g9Q==" Fecha="2018-08-10T11:30:56" Version="3.3" xmlns:cfdi="http://www.sat.gob.mx/cfd/3">
-    <cfdi:Emisor Rfc="JUEE720430NT6" Nombre="ESPERANZA JUAREZ ELIAS" RegimenFiscal="621"></cfdi:Emisor>
-    <cfdi:Receptor Rfc="KCM810226DEA" Nombre="KIMBERLY CLARK DE MEXICO, S.A.B. DE C.V." UsoCFDI="G03"></cfdi:Receptor>
-    <cfdi:Conceptos>
-        <cfdi:Concepto ClaveProdServ="90101500" Cantidad="1" ClaveUnidad="Q3" Descripcion="CONSUMO DE ALIMENTOS" ValorUnitario="215.52" Importe="215.52">
-            <cfdi:Impuestos>
-                <cfdi:Traslados>
-                    <cfdi:Traslado Base="215.52" Impuesto="002" TipoFactor="Tasa" TasaOCuota="0.160000" Importe="34.48"></cfdi:Traslado>
-                </cfdi:Traslados>
-            </cfdi:Impuestos>
-        </cfdi:Concepto>
-    </cfdi:Conceptos>
-    <cfdi:Impuestos TotalImpuestosTrasladados="34.48">
-        <cfdi:Traslados>
-            <cfdi:Traslado Impuesto="002" TipoFactor="Tasa" TasaOCuota="0.160000" Importe="34.48"></cfdi:Traslado>
-        </cfdi:Traslados>
-    </cfdi:Impuestos>
-    <cfdi:Complemento>
-        <registrofiscal:CFDIRegistroFiscal xmlns:schemaLocation="http://www.sat.gob.mx/registrofiscal http://www.sat.gob.mx/sitio_internet/cfd/cfdiregistrofiscal/cfdiregistrofiscal.xsd" Version="1.0" Folio="1808200000001981" xmlns:registrofiscal="http://www.sat.gob.mx/registrofiscal"></registrofiscal:CFDIRegistroFiscal>
-        <tfd:TimbreFiscalDigital xmlns:tfd="http://www.sat.gob.mx/TimbreFiscalDigital" xsi:schemaLocation="http://www.sat.gob.mx/TimbreFiscalDigital http://www.sat.gob.mx/sitio_internet/cfd/TimbreFiscalDigital/TimbreFiscalDigitalv11.xsd" Version="1.1" UUID="AAA100D8-2620-45E8-90D5-BCFA8E0AA708" FechaTimbrado="2018-08-10T11:32:56" RfcProvCertif="SAT970701NN3" SelloCFD="XpGru5ERb0yhJ8Fv3V64jwDAX8pJSGKEYTjCv6Y2j2jzNenmO2mntJ+93DE5dB+yTwhn5OHVZlWrEZV2OvZJ/MrP/IHdt5pZ8/ZTUs/Pm6K+BssJVUX3y20agzffLq4VzZq7zF1rxDjwN5lnsPMPlnqn0N4alC/KDSs1s1rV/GoZAh+pFtysR2a5AL17MfcjEGobu9U9MdEdHvmsc2W1ydbfnFBfVkyOqYRziHNu0m4929xxkZPrv3tse8zAZ6AxinH07YAMMQM5pv396RkqpFzsnp61M8K++wO/33XKxw1ybjRNJ4tRQ+QFL848x+8SyxTpF8vKxSNW8tg37f7g9Q==" NoCertificadoSAT="00001000000403258748" SelloSAT="M0/8XpxtQ1Bg9oLNlGhZBAAjaIpc4HGQv2DlFmaLoZFKkoeLj8yXkEK0JBPi3WaQi9yPlmGKsDfcFp1NHkZpcO98nZ+4m8EVspFyiSYTHFBu86jEszLVd19SKPE5BnMORXeuBZQGyUt1MLivzzb+7HMohHn1boUX2kgQO1wpjb52WormzNwUzYJ893zssTJgAjkhMUomdO+kCnDb9JPXv5y7qruqwX/2OTQNvU9SjLoNLKHGbSQVQkT5dEwOl3MnO8IZoIvHAB1Q9bLDICQM4oaO24bqV2uIpZjfMZuu818PROostL7qOLWLcciXtf+HdeCSa0HCKrUFP202SR5LwQ==" />
-    </cfdi:Complemento>
-</cfdi:Comprobante>
